perf(util): clean util
