+++ src/util/download.ts
@@ -3,12 +3,6 @@ import fs from "fs";
 import got from "got";
 import path from "path";
 import Progress from "progress";
-import stream from "stream";
-import tar from "tar";
-import { promisify } from "util";
-
-// constants
-const pipline = promisify(stream.pipeline);
 
 /**
  * Download the get resp to file
@@ -21,7 +15,6 @@ export async function download(
     dir: string,
     url: string,
     file: string,
-    tarFile = false,
 ): Promise<void> {
     fs.mkdirSync(dir, { recursive: true });
     const bar = new Progress(`[ ${chalk.cyan("wait")} ] [:bar] :rate/bps :etas`, {
@@ -32,22 +25,11 @@ export async function download(
     });
 
     let prePercent = 0;
-    if (tarFile) {
-        await pipline(
-            got.stream(url)
-                .on("downloadProgress", (progress) => {
-                    bar.tick(progress.percent - prePercent);
-                    prePercent = progress.percent;
-                }),
-            tar.x({ cwd: dir, strip: 1 })
-        );
-    } else {
-        const res = await got(url)
-            .on("downloadProgress", (progress) => {
-                bar.tick(progress.percent - prePercent);
-                prePercent = progress.percent;
-            });
+    const res = await got(url)
+        .on("downloadProgress", (progress) => {
+            bar.tick(progress.percent - prePercent);
+            prePercent = progress.percent;
+        });
 
-        fs.writeFileSync(path.resolve(dir, file), res.body);
-    }
+    fs.writeFileSync(path.resolve(dir, file), res.body);
 }
